tokio: deduplicate spawn_blocking
Move common code and tracing integration into Handle

Fixes #2998
Closes #3004
