+++ tokio/src/runtime/blocking/pool.rs
@@ -72,10 +72,7 @@ where
     F: FnOnce() -> R + Send + 'static,
 {
     let rt = context::current().expect("not currently running on the Tokio runtime.");
-
-    let (task, handle) = task::joinable(BlockingTask::new(func));
-    let _ = rt.blocking_spawner.spawn(task, &rt);
-    handle
+    rt.spawn_blocking(func)
 }
 
 #[allow(dead_code)]
